drf/requirements.txt.txt
Django==2.0.7
django-rest-framework==0.1.0
djangorestframework==3.8.2
Pillow==5.2.0
pytz==2018.5
whitenoise==3.3.1
dj-database-url==0.5.0